server/src/test/java/learn/game_finder/data/LocationJdbcTemplateRepositoryTest.java
package learn.game_finder.data;

import learn.game_finder.models.Location;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;

import java.util.List;

import static org.junit.jupiter.api.Assertions.*;

@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.NONE)
class LocationJdbcTemplateRepositoryTest {

    @Autowired
    LocationJdbcTemplateRepository repository;

    @Autowired
    KnownGoodState knownGoodState;

    @BeforeEach
    void setup() {
        knownGoodState.set();
    }

    @Test
    void findById() {

        Location location = repository.findById(2);
        assertNotNull(location);
        assertEquals(30.494237,location.getLatitude());

    }

    @Test
    void findAll() {
        List<Location> locations = repository.findAll();
        assertNotNull(locations);
        assertEquals(2,locations.size());

    }

    @Test
    void add() {
    }

    @Test
    void update() {
    }

    @Test
    void delete() {
    }
}